+++ cmd/cloud-controller-manager/app/BUILD
@@ -12,7 +12,6 @@ go_library(
         "//cmd/cloud-controller-manager/app/config:go_default_library",
         "//cmd/cloud-controller-manager/app/options:go_default_library",
         "//cmd/controller-manager/app:go_default_library",
-        "//pkg/controller/cloud:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/uuid:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/wait:go_default_library",
@@ -23,6 +22,7 @@ go_library(
         "//staging/src/k8s.io/client-go/tools/leaderelection/resourcelock:go_default_library",
         "//staging/src/k8s.io/cloud-provider:go_default_library",
         "//staging/src/k8s.io/cloud-provider/controllers/node:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/nodelifecycle:go_default_library",
         "//staging/src/k8s.io/cloud-provider/controllers/route:go_default_library",
         "//staging/src/k8s.io/cloud-provider/controllers/service:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
