+++ cmd/kube-controller-manager/app/BUILD
@@ -49,7 +49,6 @@ go_library(
         "//pkg/controller/certificates/rootcacertpublisher:go_default_library",
         "//pkg/controller/certificates/signer:go_default_library",
         "//pkg/controller/certificates/signer/config:go_default_library",
-        "//pkg/controller/cloud:go_default_library",
         "//pkg/controller/clusterroleaggregation:go_default_library",
         "//pkg/controller/cronjob:go_default_library",
         "//pkg/controller/daemon:go_default_library",
@@ -135,6 +134,7 @@ go_library(
         "//staging/src/k8s.io/client-go/util/flowcontrol:go_default_library",
         "//staging/src/k8s.io/client-go/util/keyutil:go_default_library",
         "//staging/src/k8s.io/cloud-provider:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/controllers/nodelifecycle:go_default_library",
         "//staging/src/k8s.io/cloud-provider/controllers/route:go_default_library",
         "//staging/src/k8s.io/cloud-provider/controllers/service:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
