+++ hack/.golint_failures
@@ -50,7 +50,6 @@ pkg/controller/apis/config/v1alpha1
 pkg/controller/certificates
 pkg/controller/certificates/signer
 pkg/controller/certificates/signer/config/v1alpha1
-pkg/controller/cloud
 pkg/controller/clusterroleaggregation
 pkg/controller/daemon
 pkg/controller/daemon/config/v1alpha1
@@ -424,6 +423,7 @@ staging/src/k8s.io/client-go/util/testing
 staging/src/k8s.io/client-go/util/workqueue
 staging/src/k8s.io/cloud-provider
 staging/src/k8s.io/cloud-provider/controllers/node
+staging/src/k8s.io/cloud-provider/controllers/nodelifecycle
 staging/src/k8s.io/cloud-provider/controllers/route
 staging/src/k8s.io/code-generator/cmd/client-gen/args
 staging/src/k8s.io/code-generator/cmd/client-gen/generators/fake
