+++ pkg/controller/BUILD
@@ -108,7 +108,6 @@ filegroup(
         "//pkg/controller/apis/config:all-srcs",
         "//pkg/controller/bootstrap:all-srcs",
         "//pkg/controller/certificates:all-srcs",
-        "//pkg/controller/cloud:all-srcs",
         "//pkg/controller/clusterroleaggregation:all-srcs",
         "//pkg/controller/cronjob:all-srcs",
         "//pkg/controller/daemon:all-srcs",
