+++ staging/src/k8s.io/cloud-provider/BUILD
@@ -38,6 +38,7 @@ filegroup(
         ":package-srcs",
         "//staging/src/k8s.io/cloud-provider/api:all-srcs",
         "//staging/src/k8s.io/cloud-provider/controllers/node:all-srcs",
+        "//staging/src/k8s.io/cloud-provider/controllers/nodelifecycle:all-srcs",
         "//staging/src/k8s.io/cloud-provider/controllers/route:all-srcs",
         "//staging/src/k8s.io/cloud-provider/controllers/service:all-srcs",
         "//staging/src/k8s.io/cloud-provider/fake:all-srcs",
