+++ staging/src/k8s.io/cloud-provider/controllers/nodelifecycle/BUILD
@@ -3,7 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["node_lifecycle_controller.go"],
-    importpath = "k8s.io/kubernetes/pkg/controller/cloud",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/controllers/nodelifecycle",
+    importpath = "k8s.io/cloud-provider/controllers/nodelifecycle",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
